refactor: remove possibility to not commit or dispose transaction
It is not a good idea to give the user of the API the possibility to decide whether a transaction is committed and disposed of. This makes the interaction far too complex. Therefore the two factory methods that have to commitTransaction and disposeTransaction variabls are removed. Instead of doing such complex life cycle management, the user now has the possibility to use the AcadDataModel and provide their database and transaction an manage the life cycle on their own.